+++ shared/src/commonMain/kotlin/org/hyperskill/app/core/injection/CommonComponent.kt
@@ -6,14 +6,14 @@ import org.hyperskill.app.core.domain.BuildVariant
 import org.hyperskill.app.core.domain.Platform
 import org.hyperskill.app.core.remote.UserAgentInfo
 import org.hyperskill.app.core.view.mapper.ResourceProvider
-import org.hyperskill.app.extension.PluralsFormatter
+import org.hyperskill.app.core.view.mapper.DateFormatter
 
 interface CommonComponent {
     val json: Json
     val userAgentInfo: UserAgentInfo
     val settings: Settings
     val resourceProvider: ResourceProvider
-    val pluralsFormatter: PluralsFormatter
+    val dateFormatter: DateFormatter
     val platform: Platform
     val buildVariant: BuildVariant
 }
