+++ shared/src/commonMain/kotlin/org/hyperskill/app/home/injection/HomeFeatureBuilder.kt
@@ -2,7 +2,7 @@ package org.hyperskill.app.home.injection
 
 import org.hyperskill.app.analytic.domain.interactor.AnalyticInteractor
 import org.hyperskill.app.core.presentation.ActionDispatcherOptions
-import org.hyperskill.app.extension.PluralsFormatter
+import org.hyperskill.app.core.view.mapper.DateFormatter
 import org.hyperskill.app.home.domain.interactor.HomeInteractor
 import org.hyperskill.app.home.presentation.HomeActionDispatcher
 import org.hyperskill.app.home.presentation.HomeFeature.Action
@@ -29,7 +29,7 @@ object HomeFeatureBuilder {
         analyticInteractor: AnalyticInteractor,
         sentryInteractor: SentryInteractor,
         urlPathProcessor: UrlPathProcessor,
-        pluralsFormatter: PluralsFormatter
+        dateFormatter: DateFormatter
     ): Feature<State, Message, Action> {
         val homeReducer = HomeReducer()
         val homeActionDispatcher = HomeActionDispatcher(
@@ -42,7 +42,7 @@ object HomeFeatureBuilder {
             analyticInteractor,
             sentryInteractor,
             urlPathProcessor,
-            pluralsFormatter
+            dateFormatter
         )
 
         return ReduxFeature(State.Idle, homeReducer)
