+++ shared/src/commonMain/kotlin/org/hyperskill/app/step_quiz/view/mapper/StepQuizStatsTextMapper.kt
@@ -1,8 +1,8 @@
 package org.hyperskill.app.step_quiz.view.mapper
 
 import org.hyperskill.app.SharedResources
+import org.hyperskill.app.core.view.mapper.DateFormatter
 import org.hyperskill.app.core.view.mapper.ResourceProvider
-import org.hyperskill.app.extension.TimeFancifier
 
 class StepQuizStatsTextMapper(
     private val resourceProvider: ResourceProvider
@@ -11,6 +11,6 @@ class StepQuizStatsTextMapper(
         resourceProvider.getString(
             SharedResources.strings.step_quiz_stats_text,
             resourceProvider.getQuantityString(SharedResources.plurals.users, users, users),
-            TimeFancifier.formatTimeDistance(millisSinceLastCompleted)
+            DateFormatter.formatTimeDistance(millisSinceLastCompleted)
         )
 }
