+++ shared/src/iosMain/kotlin/org/hyperskill/app/core/injection/CommonComponentImpl.kt
@@ -8,7 +8,7 @@ import org.hyperskill.app.core.domain.Platform
 import org.hyperskill.app.core.remote.UserAgentInfo
 import org.hyperskill.app.core.view.mapper.ResourceProvider
 import org.hyperskill.app.core.view.mapper.ResourceProviderImpl
-import org.hyperskill.app.extension.PluralsFormatter
+import org.hyperskill.app.core.view.mapper.DateFormatter
 import org.hyperskill.app.network.injection.NetworkModule
 import platform.Foundation.NSUserDefaults
 
@@ -26,8 +26,8 @@ class CommonComponentImpl(
     override val resourceProvider: ResourceProvider =
         ResourceProviderImpl()
 
-    override val pluralsFormatter: PluralsFormatter =
-        PluralsFormatter(resourceProvider)
+    override val dateFormatter: DateFormatter =
+        DateFormatter(resourceProvider)
 
     override val platform: Platform =
         Platform()
